# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AVISAUTO.fr API - Gestion d'avis Google avec IA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: List reviews for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [nouveau, proposé, modifié, publié]
        - name: sentiment
          in: query
          schema:
            type: string
            enum: [positif, neutre, négatif]
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, unreplied, replied, today, week]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsListResponse"
  /companies/{companyId}/reviews/generate-all:
    post:
      operationId: generateAllResponses
      tags: [reviews]
      summary: Generate AI responses for all new reviews
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAllRequest"
      responses:
        "200":
          description: Generated responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAllResult"
  /companies/{companyId}/reviews/publish-all:
    post:
      operationId: publishAllResponses
      tags: [reviews]
      summary: Publish all pending responses
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publish result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishAllResult"
  /companies/{companyId}/reviews/{reviewId}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a single review
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewWithResponse"
  /companies/{companyId}/reviews/{reviewId}/generate:
    post:
      operationId: generateResponse
      tags: [reviews]
      summary: Generate AI response for a review
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
  /companies/{companyId}/reviews/{reviewId}/reformulate:
    post:
      operationId: reformulateResponse
      tags: [reviews]
      summary: Reformulate response with instructions
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReformulateRequest"
      responses:
        "200":
          description: 3 response variants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReformulateResult"
  /companies/{companyId}/reviews/{reviewId}/save-response:
    post:
      operationId: saveResponse
      tags: [reviews]
      summary: Save final response text
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveResponseRequest"
      responses:
        "200":
          description: Response saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
  /companies/{companyId}/reviews/{reviewId}/publish:
    post:
      operationId: publishResponse
      tags: [reviews]
      summary: Publish response to Google
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
  /companies/{companyId}/reviews/{reviewId}/mark-replied:
    post:
      operationId: markReplied
      tags: [reviews]
      summary: Mark review as replied manually
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as replied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /companies/{companyId}/subscription:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription details
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /companies/{companyId}/subscription/checkout:
    post:
      operationId: createCheckout
      tags: [subscriptions]
      summary: Create Stripe checkout session
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Review:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
          nullable: true
        publishDate:
          type: string
          nullable: true
        ownerReply:
          type: string
          nullable: true
        status:
          type: string
          enum: [nouveau, proposé, modifié, publié]
        sentiment:
          type: string
          enum: [positif, neutre, négatif]
          nullable: true
        createdAt:
          type: string
          format: date-time
        aiResponse:
          $ref: "#/components/schemas/AiResponse"
          nullable: true
      required:
        - id
        - companyId
        - authorName
        - rating
        - status
        - createdAt
    ReviewWithResponse:
      allOf:
        - $ref: "#/components/schemas/Review"
    ReviewsListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - reviews
        - total
        - page
        - totalPages
    AiResponse:
      type: object
      properties:
        id:
          type: integer
        reviewId:
          type: integer
        companyId:
          type: integer
        generatedText:
          type: string
        finalText:
          type: string
          nullable: true
        tone:
          type: string
          enum: [chaleureux, professionnel, direct]
        length:
          type: string
          enum: [court, moyen, long]
        style:
          type: string
          enum: [remerciement, excuse, résolution, neutre]
        status:
          type: string
          enum: [proposé, en_attente_publication, publié]
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reviewId
        - companyId
        - generatedText
        - tone
        - length
        - style
        - status
        - createdAt
...
    GenerateRequest:
      type: object
      properties:
        tone:
          type: string
          enum: [chaleureux, professionnel, direct]
        length:
          type: string
          enum: [court, moyen, long]
        style:
          type: string
          enum: [remerciement, excuse, résolution, neutre]
      required:
        - tone
        - length
        - style
    GenerateAllRequest:
      type: object
      properties:
        tone:
          type: string
          enum: [chaleureux, professionnel, direct]
        length:
          type: string
          enum: [court, moyen, long]
        style:
          type: string
          enum: [remerciement, excuse, résolution, neutre]
    GenerateAllResult:
      type: object
      properties:
        generated:
          type: integer
        total:
          type: integer
        message:
          type: string
      required:
        - generated
        - total
        - message
    ReformulateRequest:
      type: object
      properties:
        instructions:
          type: string
        currentText:
          type: string
        tone:
          type: string
          enum: [chaleureux, professionnel, direct]
        length:
          type: string
          enum: [court, moyen, long]
      required:
        - instructions
        - currentText
    ReformulateResult:
      type: object
      properties:
        variants:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 3
      required:
        - variants
    SaveResponseRequest:
      type: object
      properties:
        finalText:
          type: string
        status:
          type: string
          enum: [proposé, en_attente_publication, publié]
      required:
        - finalText
    PublishResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        publishedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - success
        - message
    PublishAllResult:
      type: object
      properties:
        published:
          type: integer
        total:
          type: integer
        message:
          type: string
        timeSaved:
          type: string
      required:
        - published
        - total
        - message
        - timeSaved
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        stripeCustomerId:
          type: string
          nullable: true
        stripeSubscriptionId:
          type: string
          nullable: true
        plan:
          type: string
          enum: [starter, pro, agency]
          nullable: true
        status:
          type: string
          enum: [active, past_due, canceled, none]
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
      required:
        - companyId
        - status
    CheckoutRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [starter, pro, agency]
        billingPeriod:
          type: string
          enum: [monthly, annual]
      required:
        - plan
        - billingPeriod
    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...