# lib/api-zod/src/generated/types/reformulateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AVISAUTO.fr API - Gestion d'avis Google avec IA
 * OpenAPI spec version: 0.1.0
 */

export interface ReformulateResult {
  /**
   * @minItems 3
   * @maxItems 3
   */
  variants: string[];
}